refactor: reorganize toast management by introducing a new ToastHelper class, enhancing toast notification capabilities, and removing legacy toast service for improved clarity and maintainability
