+++ packages/masterfabric-expo-core/src/helpers/index.ts
@@ -16,6 +16,7 @@ export * from './string_helper';
 
 // UI Feedback Helpers
 export * from './snackbar_helper';
+export * from './toast_helper';
 
 // Re-export types with different names to avoid conflicts
 export type { DeviceInfo as DeviceInfoHelper, DeviceInfoOptions } from './device-info';
