--- packages/masterfabric-expo-core/src/types/Toast.ts
@@ -1,1 +1,0 @@
-
