--- src/screens/toast-helper/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './use-toast-helper-view-model';
