--- src/screens/toast-helper/models/index.ts
@@ -1,1 +1,0 @@
-export * from './toast-helper.models';
