--- src/screens/toast-helper/store/index.ts
@@ -1,1 +1,0 @@
-export * from './toast-helper-store';
