--- src/screens/toast-helper/styles/index.ts
@@ -1,6 +1,0 @@
-export * from './toast-container.styles';
-export * from './toast-helper-screen.styles';
-export * from './toast-input-field.styles';
-export * from './toast-result-card.styles';
-export * from './toast.styles';
-
